% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_functions.R
\name{add_video}
\alias{add_video}
\title{add_video Add video to specific playlist_id}
\usage{
add_video(url, query, body, config)
}
\arguments{
\item{url}{Google specific url for this type of api call.}
}
\value{

}
\description{
add_video Add video to specific playlist_id
}
