% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{get_batch_videos}
\alias{get_batch_videos}
\title{---------
get_batch_videos
---------}
\usage{
get_batch_videos(base_url, query, config)
}
\arguments{
\item{base_url}{ur_provide_by_api}

\item{query}{query_string}

\item{config}{good_question}
}
\value{

}
\description{
---------
get_batch_videos
---------
}
