% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{get_json}
\alias{get_json}
\title{----------------
function get_json
-----------------}
\usage{
get_json(r)
}
\value{
json_content
}
\description{
----------------
function get_json
-----------------
}
