% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{get_nextPageToken}
\alias{get_nextPageToken}
\title{----------------
function nextPageToken
----------------}
\usage{
get_nextPageToken(r)
}
\arguments{
\item{r}{response}
}
\value{

}
\description{
----------------
function nextPageToken
----------------
}
