% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{get_typical_yt}
\alias{get_typical_yt}
\title{function get_typical_yt
DECRECATED - do not use}
\usage{
get_typical_yt()
}
\value{

}
\description{
function get_typical_yt
DECRECATED - do not use
}
