% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_functions.R
\name{set_body}
\alias{set_body}
\title{helper functions to add video to playlist}
\usage{
set_body(video_id, playlist_id)
}
\arguments{
\item{video_id}{The youtube assigned id.}

\item{playlist_id}{The youtube assigned playlist id.}
}
\value{
The body portion (header) of the call to the API.
}
\description{
helper function: set_body 
prepare data portion of query to yt
}
\section{Section 1.}{
NA
}

\section{Section 2.}{
NA
}

\seealso{

}
