% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{set_query}
\alias{set_query}
\title{---------
set_query
---------}
\usage{
set_query(part, maxResults = 50, mine = NULL, fields = NULL, key, ...)
}
\arguments{
\item{part}{defined_by_api.}

\item{maxResults}{number.}

\item{mine}{my_info.}

\item{fields}{api_specific}

\item{key}{api_specific}
}
\value{
list
}
\description{
sets query portion of POST.
  * item1
  * item2

Additional paragraph.   As much as you want.
}
\examples{
\dontrun{
set_query("......")
}
}
