% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{get_batch_videos}
\alias{get_batch_videos}
\title{---------
get_batch_videos
---------}
\usage{
get_batch_videos(base_url, query, config)
}
\value{

}
\description{
---------
get_batch_videos
---------
}
