% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_api_functions.R
\name{get_typical_yt}
\alias{get_typical_yt}
\title{-----------------
get_typical_yt
-----------------
returns pre-set examples for channelId, playlistId, videoId
  and maxResults = 50
playlistId <- "PLbcglKxZP5PMBoG1VfCqlWZ1Nx4FBBOn6"  # 60s - female (300+)
videoId <- "UCVdHFHpFBg"   # Tex Ritter Range Party}
\usage{
get_typical_yt()
}
\value{
list
}
\description{
-----------------
get_typical_yt
-----------------
returns pre-set examples for channelId, playlistId, videoId
  and maxResults = 50
playlistId <- "PLbcglKxZP5PMBoG1VfCqlWZ1Nx4FBBOn6"  # 60s - female (300+)
videoId <- "UCVdHFHpFBg"   # Tex Ritter Range Party
}
