% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_functions.R
\name{set_body}
\alias{set_body}
\title{helper functions to add video to playlist
===========
set_body
===========
helper function: set_body 
prepare data portion of query to yt}
\usage{
set_body(video_id, playlist_id)
}
\value{

}
\description{
helper functions to add video to playlist
===========
set_body
===========
helper function: set_body 
prepare data portion of query to yt
}
